Tyr: move aggregatePlaceCommand in a "command" submodule
