+++ source/tyr/tyr/aggregate_places.py
@@ -7,7 +7,6 @@ from geoalchemy2.types import Geography
 from geoalchemy2.shape import to_shape
 from sqlalchemy.ext.declarative import declarative_base
 import logging
-from flask.ext.script import Command, Option
 from tyr.helper import load_instance_config
 
 def make_sqlalchemy_connection_string(instance_config):
@@ -150,35 +149,4 @@ def aggregate_places(instance_config, job_id):
         handle_object_instance("network", NetworkEd)
     finally:
         session_ed.close()
-
-
-class AggregatePlacesCommand(Command):
-#A command used in development environment to run aggregate places without
-#having to run tyr
-
-    def get_options(self):
-        return [
-            Option('-n', '--name', dest='name_', default="",
-                   help="Launch aggregate places, if no name is specified "
-                   "this is launched for all instances")
-        ]
-
-    def run(self, name_=""):
-        if name_:
-            instances = models.Instance.query.filter_by(name=name_).all()
-        else:
-            instances = models.Instance.query.all()
-
-        job_id = 1
-        instances_name = [instance.name for instance in instances]
-        for instance_name in instances_name:
-            instance_config = None
-            try:
-                instance_config = load_instance_config(instance_name)
-            except ValueError:
-                logging.getLogger(__name__).\
-                    info("Unable to find instance " + instance_name)
-                continue
-            aggregate_places(instance_config, job_id)
-            job_id += 1
